+++ src/test/java/com/floney/floney/config/QueryDslTest.java
@@ -1,0 +1,11 @@
+package com.floney.floney.config;
+
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
+
+@Import(QueryFactoryConfig.class)
+@DataJpaTest
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+public @interface QueryDslTest {
+}
