--- src/test/java/com/floney/floney/config/TestConfig.java
@@ -1,20 +1,0 @@
-package com.floney.floney.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-@TestConfiguration
-public class TestConfig {
-
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory queryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
